Move EmailHtmlBody to notifications
